LPS-57070 move class - same pattern as application-list (no changes)
