+++ modules/apps/search/search-api/src/com/liferay/search/api/facet/BaseSearchFacet.java
@@ -23,7 +23,6 @@ import com.liferay.portal.kernel.search.facet.Facet;
 import com.liferay.portal.kernel.search.facet.config.FacetConfiguration;
 import com.liferay.portal.kernel.search.facet.util.FacetFactoryUtil;
 import com.liferay.portal.kernel.util.Validator;
-import com.liferay.search.api.util.SearchFacet;
 
 /**
  * @author Eudaldo Alonso
