+++ modules/apps/search/search-web/src/com/liferay/search/web/facet/ModifiedSearchFacet.java
@@ -23,7 +23,7 @@ import com.liferay.portal.kernel.search.facet.config.FacetConfiguration;
 import com.liferay.portal.kernel.util.ParamUtil;
 import com.liferay.portal.kernel.util.StringUtil;
 import com.liferay.search.api.facet.BaseJSPSearchFacet;
-import com.liferay.search.api.util.SearchFacet;
+import com.liferay.search.api.facet.SearchFacet;
 
 import javax.portlet.ActionRequest;
