+++ modules/apps/search/search-web/src/com/liferay/search/web/facet/UserSearchFacet.java
@@ -21,7 +21,7 @@ import com.liferay.portal.kernel.search.facet.MultiValueFacet;
 import com.liferay.portal.kernel.search.facet.config.FacetConfiguration;
 import com.liferay.portal.kernel.util.ParamUtil;
 import com.liferay.search.api.facet.BaseJSPSearchFacet;
-import com.liferay.search.api.util.SearchFacet;
+import com.liferay.search.api.facet.SearchFacet;
 
 import javax.portlet.ActionRequest;
